Wire up crash reproducers.
Uses the IREE API `ireeCompilerInvocationSetCrashHandler` from https://github.com/iree-org/iree/pull/11845.

Mostly this is a re-organization of some classes into debugging.h to get layering right (so that the compiler can depend on the artifact dumper).

Testing this is a bit hard, but I expect it will be getting plenty of use in the next step (running the Jax test suite).
